Moving utiltiy type to common folder
